Update IBenchHelper.java
Unless I am missing something, the extra check for any iBench in the player's inventory only serves to link up the inventories of any iBenches the player has (similar to an ender-chest). If you want it to act this way, disregard my pull request. But if you want each iBench instance/ItemStack to have it's own inventory, deleting that code does the trick.